T H I C C C updates - ES6 compliant
