+++ src/bot.ts
@@ -1,51 +1,45 @@
-import HTTPS = require('https');
-import express = require('express');
+import { Response } from 'express';
 import { ClientRequest } from 'http';
-import { inject, injectable } from 'inversify';
-import CommandFactory from './commands/commandFactory';
+import { request, RequestOptions } from 'https';
+import { CommandFactory } from './commands/commandFactory';
 import { SenderType } from './constants/GroupmeSenderType';
 import { BotResponseModel } from './models/BotResponseModel';
-import GroupmeMessageModel from './models/Groupme/GroupmeMessageModel';
+import { GroupmeMessageModel } from './models/Groupme/GroupmeMessageModel';
 
-@injectable()
-export default class Bot {
+export class Bot {
 
-  commandFactory: CommandFactory;
+  private static readonly GROUPME_HOSTNAME: string = 'api.groupme.com';
+  private static readonly GROUPME_PATH: string = '/v3/bots/post';
 
-  constructor(@inject(CommandFactory) cmdFactory: CommandFactory) {
-    this.commandFactory = cmdFactory;
-  }
-
-  async respond(reqBody: GroupmeMessageModel, response: express.Response): Promise<void> {
-    if (reqBody.sender_type === SenderType.Bot) return Promise.resolve();
-    if (reqBody.text) reqBody.text = reqBody.text.trim().toLowerCase();
+  static async respond(reqBody: GroupmeMessageModel, response: Response): Promise<void> {
+    if (reqBody.sender_type === SenderType.Bot || !reqBody.text) {
+      response.writeHead(200);
+      response.end('Not responding to nonsense');
+      return Promise.resolve(); // Don't respond to other bots or empty images
+    }
 
     let responseMsg: string;
-    response.writeHead(200);
-    const command = this.commandFactory.getCommand(reqBody.text);
+    let cleanText = reqBody.text.trim().toLowerCase();
+    const command = CommandFactory.getCommand(cleanText);
     if (command) {
-      const results = await command.execute(reqBody);
-      responseMsg = results.text;
-      this.sendBotResponse(results);
+      try {
+        const results = await command.execute(reqBody);
+        response.writeHead(200);
+        responseMsg = results.text;
+        this.sendBotResponse(results);
+      }
+      catch{
+        response.writeHead(500);
+        responseMsg = `Error executing: ${cleanText}`;
+      }
+    } else {
+      responseMsg = 'No command found';
     }
 
     response.end(responseMsg);
   }
 
-  private sendBotResponse(responseModel: BotResponseModel) {
-    const reqOptions: HTTPS.RequestOptions = {
-      hostname: 'api.groupme.com',
-      path: '/v3/bots/post',
-      method: 'POST'
-    };
-
-    const clientReq = HTTPS.request(reqOptions);
-    this.addRequestErrorHandlers(clientReq);
-    console.log(`This is what I'm sending back: "${responseModel.text}"`);
-    clientReq.end(JSON.stringify(responseModel));
-  };
-
-  private addRequestErrorHandlers(clientReq: ClientRequest) {
+  private static addRequestErrorHandlers(clientReq: ClientRequest) {
     clientReq.on('error', function (err: any) {
       console.log('error posting message ' + JSON.stringify(err));
     });
@@ -53,5 +47,17 @@ export default class Bot {
       console.log('timeout posting message ' + JSON.stringify(err));
     });
   }
-}
 
+  private static sendBotResponse(responseModel: BotResponseModel) {
+    const reqOptions: RequestOptions = {
+      hostname: this.GROUPME_HOSTNAME,
+      path: this.GROUPME_PATH,
+      method: 'POST'
+    };
+
+    const clientReq = request(reqOptions);
+    this.addRequestErrorHandlers(clientReq);
+    console.log(`Sending: "${responseModel.text}"`);
+    clientReq.end(JSON.stringify(responseModel));
+  };
+}
