+++ src/commands/constants/commandList.ts
@@ -1,5 +1,4 @@
 import { AwardsCommand } from "../models/awardsCommand"
-import { Command } from "../models/command"
 import { CoolGuyCommand } from "../models/coolGuyCommand"
 import { HelpCommand } from "../models/helpCommand"
 import { HereCommand } from "../models/hereCommand"
@@ -7,12 +6,14 @@ import { SpewCommand } from "../models/spewCommand"
 import { ThesaurizeCommand } from "../models/thesaurizeCommand"
 import { WednesdayCommand } from "../models/wednesdayCommand"
 
-export default [
-  AwardsCommand,
-  CoolGuyCommand,
-  HelpCommand,
-  HereCommand,
-  SpewCommand,
-  ThesaurizeCommand,
-  WednesdayCommand
-] as unknown as Command[]
+export class CommandList{
+  public static readonly allCommands = [
+    AwardsCommand,
+    CoolGuyCommand,
+    HelpCommand,
+    HereCommand,
+    SpewCommand,
+    ThesaurizeCommand,
+    WednesdayCommand
+  ];
+}
