+++ src/commands/models/awardsCommand.spec.ts
@@ -1,11 +1,10 @@
+import sinon from 'sinon';
+import { AwardsTool } from '../../tools/awardsTool';
 import { AwardsCommand } from './awardsCommand';
-import sinon = require('sinon');
-
-import AwardsService = require('../../services/awardSvc');
 
 describe('Awards Command Tests', () => {
   it('should have the awardSvc.getAwards method run on execute', () => {
-    let awardSvcSpy = sinon.stub(AwardsService, 'getAwards').returns(Promise.resolve('Test Awards Results'));
+    let awardSvcSpy = sinon.stub(AwardsTool, 'getAwards').returns(Promise.resolve('Test Awards Results'));
     let awardsCommand = new AwardsCommand();
 
     awardsCommand.execute();
