+++ src/commands/models/spewCommand.ts
@@ -1,9 +1,9 @@
 import { BotResponseModel } from "../../models/BotResponseModel";
-import GroupmeMessageModel from "../../models/Groupme/GroupmeMessageModel";
+import { GroupmeMessageModel } from "../../models/Groupme/GroupmeMessageModel";
+import { GroupmeTool } from "../../tools/groupmeTool";
 import { CommandCheckLocation } from "../constants/commandCheckLocation";
 import { Command } from "./command";
 
-import GroupMeService = require('../../services/gmeSvc');
 
 export class SpewCommand implements Command {
 
@@ -14,7 +14,7 @@ export class SpewCommand implements Command {
   constructor() { }
 
   async execute(botRequestBody: GroupmeMessageModel): Promise<BotResponseModel> {
-    const result = await GroupMeService.addSpew(parseInt(botRequestBody.user_id));
+    const result = await GroupmeTool.addSpew(parseInt(botRequestBody.user_id));
     return new BotResponseModel(result, null);
   }
 }
