+++ src/commands/models/wednesdayCommand.ts
@@ -1,4 +1,4 @@
-import moment = require('moment');
+import moment from 'moment';
 import { BotResponseModel } from "../../models/BotResponseModel";
 import { CommandCheckLocation } from '../constants/commandCheckLocation';
 import { Command } from "./command";
