+++ src/models/BotResponseModel.ts
@@ -1,10 +1,10 @@
 import { GroupmeAttachmentModel } from "./Groupme/GroupmeAttachmentModel";
 
 export class BotResponseModel {
-  text: string;
-  attachments: GroupmeAttachmentModel[];
+  public text: string;
+  public attachments: GroupmeAttachmentModel[];
 
-  bot_id: string;
+  public bot_id: string;
 
   constructor(text: string, attachments: GroupmeAttachmentModel[]) {
     this.bot_id = process.env.BOT_ID;
