+++ tsconfig.json
@@ -12,7 +12,8 @@
       "types": ["reflect-metadata"],
       "moduleResolution": "node",
       "experimentalDecorators": true,
-      "emitDecoratorMetadata": true
+      "emitDecoratorMetadata": true,
+      "esModuleInterop": true
     },
     "include": [
         "./src/**/*"
